+++ include/libfork/core/impl/awaitables.hpp
@@ -21,8 +21,8 @@
 #include "libfork/core/co_alloc.hpp"          // for co_allocable, co_new_t, stack_allocated
 #include "libfork/core/exception.hpp"         // for exception_before_join
 #include "libfork/core/ext/context.hpp"       // for full_context
-#include "libfork/core/ext/handles.hpp"       // for submit_handle, submit_t, task_handle
-#include "libfork/core/ext/list.hpp"          // for unwrap, intrusive_list
+#include "libfork/core/ext/handles.hpp"       // for submit_handle, submit_node_t, task_handle
+#include "libfork/core/ext/list.hpp"          // for unwrap
 #include "libfork/core/ext/tls.hpp"           // for stack, context
 #include "libfork/core/impl/frame.hpp"        // for frame
 #include "libfork/core/impl/stack.hpp"        // for stack
@@ -170,7 +170,7 @@ struct context_switch_awaitable {
   /**
    * @brief The current coroutine's handle.
    */
-  intrusive_list<impl::submit_t *>::node self;
+  submit_node_t self;
 };
 
 // -------------------------------------------------------- //