+++ include/libfork/core/sync_wait.hpp
@@ -11,22 +11,25 @@
 
 #include <bit>         // for bit_cast
 #include <exception>   // for rethrow_exception
-#include <optional>    // for optional, nullopt
+#include <memory>      // for make_shared
+#include <optional>    // for optional
 #include <semaphore>   // for binary_semaphore
 #include <type_traits> // for conditional_t
 #include <utility>     // for forward
 
-#include "libfork/core/eventually.hpp"           // for basic_eventually
-#include "libfork/core/ext/handles.hpp"          // for submit_t
+#include "libfork/core/defer.hpp"                // for LF_DEFER
+#include "libfork/core/eventually.hpp"           // for try_eventually
+#include "libfork/core/exception.hpp"            // for sync_wait_in_worker
+#include "libfork/core/ext/handles.hpp"          // for submit_node_t, submit_t
 #include "libfork/core/ext/list.hpp"             // for intrusive_list
-#include "libfork/core/ext/tls.hpp"              // for thread_stack, has_stack
+#include "libfork/core/ext/tls.hpp"              // for has_stack, thread_stack, has_context
 #include "libfork/core/first_arg.hpp"            // for async_function_object
 #include "libfork/core/impl/combinate.hpp"       // for quasi_awaitable, y_combinate
 #include "libfork/core/impl/frame.hpp"           // for frame
 #include "libfork/core/impl/manual_lifetime.hpp" // for manual_lifetime
-#include "libfork/core/impl/stack.hpp"           // for stack, swap
+#include "libfork/core/impl/stack.hpp"           // for stack
 #include "libfork/core/invocable.hpp"            // for async_result_t, ignore_t, rootable
-#include "libfork/core/macro.hpp"                // for LF_LOG, LF_CLANG_TLS_NOINLINE
+#include "libfork/core/macro.hpp"                // for LF_CLANG_TLS_NOINLINE
 #include "libfork/core/scheduler.hpp"            // for scheduler
 #include "libfork/core/tag.hpp"                  // for tag, none
 
@@ -84,9 +87,9 @@ inline namespace core {
  * is expected to make a call from `main` into a scheduler/runtime by scheduling a single root-task with this
  * function.
  *
- * This will build a task from `fun` and dispatch it to `sch` via its `schedule` method. Sync wait should
- * __not__ be called by a worker thread (which are never allowed to block) unless the call to `schedule`
- * completes synchronously.
+ * This will build a task from `fun` and dispatch it to `sch` via its `schedule` method. If `sync_wait` is
+ * called by a worker thread (which are never allowed to block) then `lf::core::sync_wait_in_worker` will be
+ * thrown.
  */
 template <scheduler Sch, async_function_object F, class... Args>
   requires rootable<F, Args...>
@@ -94,60 +97,64 @@ LF_CLANG_TLS_NOINLINE auto sync_wait(Sch &&sch, F fun, Args &&...args) -> async_
 
   std::binary_semaphore sem{0};
 
-  // This is to support a worker sync waiting on work they will launch inline.
-  bool worker = impl::tls::has_stack;
-  // Will cache workers stack here.
-  std::optional<impl::stack> prev = std::nullopt;
-
-  basic_eventually<async_result_t<F, Args...>, true> result;
-
-  impl::y_combinate combinator = combinate<tag::root, modifier::none>(&result, std::move(fun));
-
-  if (!worker) {
-    LF_LOG("Sync wait from non-worker thread");
-    impl::tls::thread_stack.construct();
-    impl::tls::has_stack = true;
-  } else {
-    LF_LOG("Sync wait from worker thread");
-    prev.emplace();                        // Default construct.
-    swap(*prev, *impl::tls::thread_stack); // ADL call.
+  if (impl::tls::has_stack || impl::tls::has_context) {
+    throw sync_wait_in_worker{};
   }
 
+  // Initialize the non-workers stack.
+  impl::tls::thread_stack.construct();
+  impl::tls::has_stack = true;
+
+  // Clean up the stack on exit.
+  LF_DEFER {
+    impl::tls::thread_stack.destroy();
+    impl::tls::has_stack = false;
+  };
+
+  using eventually_t = try_eventually<async_result_t<F, Args...>>;
+
+  // If we fail to wait for the result to complete due to an exception we will need
+  // to detach the coroutine, this will require the node and the result to be stored
+  // on the heap such that returning from this function will not destroy them.
+  struct heap_alloc : eventually_t {
+
+    using eventually_t::operator=;
+
+    auto operator=(heap_alloc &&other) -> heap_alloc & = delete;
+    auto operator=(heap_alloc const &other) -> heap_alloc & = delete;
+
+    std::optional<impl::submit_node_t> node; // Make default constructible.
+  };
+
+  auto heap = std::make_shared<heap_alloc>();
+
+  // Build a combinator, copies heap shared_ptr.
+  impl::y_combinate combinator = combinate<tag::root, modifier::none>(heap, std::move(fun));
   // This allocates a coroutine on this threads stack.
   impl::quasi_awaitable await = std::move(combinator)(std::forward<Args>(args)...);
-
+  // Set the root semaphore.
   await->set_root_sem(&sem);
 
-  auto *handle = std::bit_cast<impl::submit_t *>(await.release()); // Ownership transferred!
+  // If this throws then `await` will clean up the coroutine.
+  impl::ignore_t{} = impl::tls::thread_stack->release();
 
-  // TODO: this could be made exception safe.
+  // We will pass a pointer to this to schedule.
+  heap->node.emplace(std::bit_cast<impl::submit_t *>(await.get()));
 
-  [&]() noexcept {
-    // If this threw we could clean up coroutine but we would need to repair the worker state.
-    impl::ignore_t{} = impl::tls::thread_stack->release();
+  // Schedule upholds the strong exception guarantee hence, if it throws `await` cleans up.
+  std::forward<Sch>(sch).schedule(&*heap->node);
+  // If -^ didn't throw then we release ownership of the coroutine.
+  impl::ignore_t{} = await.release();
 
-    if (!worker) {
-      impl::tls::thread_stack.destroy();
-      impl::tls::has_stack = false;
-    } else {
-      swap(*prev, *impl::tls::thread_stack);
-    }
+  // If this throws that's ok as `result` and `node` are on the heap.
+  sem.acquire();
 
-    typename intrusive_list<impl::submit_t *>::node node{handle};
-
-    // If this threw we could clean up the coroutine if we repaired the worker state.
-    std::forward<Sch>(sch).schedule(&node);
-
-    // If this threw we would have to terminate, unless the return variable was stored on the heap.
-    sem.acquire();
-  }();
-
-  if (result.has_exception()) {
-    std::rethrow_exception(std::move(result).exception());
+  if (heap->has_exception()) {
+    std::rethrow_exception(std::move(*heap).exception());
   }
 
   if constexpr (!std::is_void_v<async_result_t<F, Args...>>) {
-    return *std::move(result);
+    return *std::move(*heap);
   }
 }